+++ .gitignore
@@ -10,3 +10,4 @@ a.out
 .vscode
 sample.c
 ls
+redirect
